Avoid unrelated exception in tracebacks (#32)
* Avoid unrelated exception in tracebacks

In the case there is no async library, current_async_library will raise an exception. We then go on to call asyncio_run while handling the exception. The consequence is that, if the asyncio_run call itself raises, the current_async_library exception ends up on the traceback. This unrelated exception then ends up cluttering the traceback.

* Fix installed python version

There is no matrix on these steps